Move employee-reads-application tests to Playwright
